+++ CustomerManagementSystem/CustomerManagementSystem.csproj
@@ -205,7 +205,7 @@
     <Compile Include="Models\RepositoryHelper.cs">
       <DependentUpon>IRepository.tt</DependentUpon>
     </Compile>
-    <Compile Include="ViewModels\Out\CustomersQueryViewModel.cs" />
+    <Compile Include="ViewModels\CustomersQueryViewModel.cs" />
     <Compile Include="Models\客戶聯絡人.cs">
       <DependentUpon>客戶資料Model.tt</DependentUpon>
     </Compile>
@@ -250,8 +250,8 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
-    <Compile Include="ViewModels\In\CustomerQueryInModel.cs" />
-    <Compile Include="ViewModels\Out\CustomersViewModel.cs" />
+    <Compile Include="ViewModels\CustomerQueryInModel.cs" />
+    <Compile Include="ViewModels\CustomersViewModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Content\bootstrap.css" />
