+++ CustomerManagementSystem/Models/客戶資料Repository.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Linq;
 using System.Collections.Generic;
-using CustomerManagementSystem.ViewModels.In;
+using CustomerManagementSystem.ViewModels;
 
 namespace CustomerManagementSystem.Models
 {
